pyproject.toml
[tool.poetry]
name = "deep-ancestry"
version = "0.1.0"
description = ""
authors = ["Aleksandr Medvedev <alex@genx.global>"]
readme = "README.md"
packages = [{include = "deep_ancestry"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
numpy = "^1.24.3"
flwr = "^1.4.0"
pandas = "^2.0.1"
tqdm = "^4.65.0"
scikit-learn = "^1.2.2"
pytorch-lightning = "^2.0.2"
mlflow = "^2.3.1"
fire = "^0.5.0"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# my_package_cli = 'my_package.console:run'
ancestry = "scripts:global_ancestry"
plink2 = "scripts:plink2"